+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Last changes:
 
 
+Version `3.3.2a6 (20 Mar'21)`
+
+  - CS rearranged & renamed
+  - active/focus colors for sca/sb*/nbk.tab
+  - "-after/-before %w" for packing sbv/sbh instead of default setting
+  - "-minsize" for showModal/showWindow
+  - "-pointer/pointer+x+y" for showModal/showWindow's -geometry
+  - "-escape" option for showModal
+  - "-wrap word" default for 'tex'
+  - writeTextFile & error procs introduced
+  - defaultAttrs instead of setDefaultAttrs, with more options
+  - icons for mc-labels
+  - "pack/grid propagate" for "pan" if -w and -h set
+  - clrpick.tcl updated
+  - bartabs updated
+  - BUGFIX: menu's colors restored for themes
+
+
 Version `3.3.1 (6 Mar'21)`
 
   - better handling for modal/non-modal windows